// src/app/modules/ci/transacciones/paquete/model/documentoInvDet-api-model-interface.ts
export interface ICrearDocumnetoInvDet {
	documentoInvEncId: number;
	lineadocinv: number;
	tipo: string;
	subtipo: string;
	subsubtipo: string;
	cantidad: number;
	costototallocal: number;
	costototaldolar: number;
	preciototallocal: number;
	preciototaldolar: number;
	consecutivoId: number;
	ajusteconfigId: number;
	articuloid: number;
	bodegadestinoid: number;
	bodegaid: number;
	centrocuentaid: number;
	locdestinoid: number;
	localizacionid: number;
	loteId: number;
	paqueteinventarioid: number;
	tipoPagoId: number;
}