// src/app/modules/as/tablas/otros/bodega/service/existenciaBodega-api-model-interface.ts
export interface IResponseCreateExistenciaBodega {
	articuloId: number;
	bodegaId: number;
	existenciaMinima: number;
	existenciaMaxima: number;
	puntoDeOrden: number;
}

export interface IResponseConsultarExistenciaBodega {
	id: number;
	articuloId: number;
	bodegaId: number;
	existenciaMinima: number;
	existenciaMaxima: number;
	puntoDeOrden: number;
	cantDisponible: number;
	cantReservada: number;
	cantNoAprobada: number;
	cantVencida: number;
	cantTransito: number;
	cantProduccion: number;
	cantPedida: number;
	cantRemitida: number;
	congelado: boolean;
	fechaCong: string;
	bloqueaTrans: boolean;
	fechaDescong: string;
	costoUntPromedioLoc: number;
	costoUntPromedioDol: number;
}

export interface IResponseConsultarArticuloBodega {
	id: number;
	articuloId: number;
	bodegaId: number;
	existenciaMinima: number;
	existenciaMaxima: number;
	puntoDeOrden: number;
	cantDisponible: number;
	cantReservada: number;
	cantNoAprobada: number;
	cantVencida: number;
	cantTransito: number;
	cantProduccion: number;
	cantPedida: number;
	cantRemitida: number;
	congelado: boolean;
	fechaCong: string;
	bloqueaTrans: boolean;
	fechaDescong: string;
	costoUntPromedioLoc: number;
	costoUntPromedioDol: number;
	articulo: {
		activo: boolean;
		articuloDelProv: string;
		bultos: number;
		calculaPercep: boolean;
		categoriaArticuloId: number;
		claseABC: string;
		clasificacion1: string;
		clasificacion1Id: number;
		clasificacion2: string;
		clasificacion2Id: number;
		clasificacion3: string;
		clasificacion3Id: number;
		clasificacion4: string;
		clasificacion4Id: number;
		clasificacion5: string;
		clasificacion5Id: number;
		clasificacion6: string;
		clasificacion6Id: number;
		codArticulo: string;
		codigoBarrasInvt: string;
		codigoBarrasVent: string;
		codigoRetencion: string;
		color: string;
		costoDol: number;
		costoLoc: number;
		costoPromUltimoDol: number;
		costoPromUltimoLoc: number;
		createDate: string;
		createdby: string;
		descripcion: string;
		diasCuarentena: number;
		esImpuesto: boolean;
		estilo: string;
		existenciaMaxima: number;
		existenciaMinima: number;
		factorConver1: number;
		factorConver2: number;
		factorConver3: number;
		factorConver4: number;
		factorConver5: number;
		factorConver6: number;
		factorEmpaque: number;
		factorVenta: number;
		fechaHoraCreacion: string;
		fechaHoraUltModif: string;
		frecuenciaConteo: number;
		id: number;
		impuestoId: number;
		isDeleted: boolean;
		loteMultiplo: number;
		manufacturador: string;
		minVidaCompra: number;
		minVidaConsumo: number;
		minVidaVenta: number;
		modalidadInvFis: string;
		modeloRetencion: string;
		notas: string;
		obligaCuarentena: boolean;
		ordenMinima: number;
		perecedero: boolean;
		pesoBruto: number;
		pesoNeto: number;
		plazoReabast: number;
		porcPercep: number;
		precioBaseDol: number;
		precioBaseLocal: number;
		proveedorId: number;
		puntoDeOrden: number;
		retencionCompra: string;
		retencionVenta: string;
		sugiereMin: boolean;
		talla: string;
		tiendaEnLinea: boolean;
		tipo: string;
		tipoCodBarraAlm: string;
		tipoCodBarraDet: string;
		tipoCosto: string;
		tipoDocIVA: string;
		ultimaSalida: string;
		ultimoIngreso: string;
		ultimoInventario: string;
		ultimoMovimiento: string;
		unidadAlmacenId: number;
		unidadEmpaqueId: number;
		unidadVentaId: number;
		updateDate: string;
		updatedby: string;
		urlimagen: string;
		usaLotes: boolean;
		usaNumerosSerie: boolean;
		usaReglasLocales: boolean;
		usuarioCreacion: string;
		usuarioUltModif: string;
		ventaTarjetaCredito: boolean;
		vidaUtilPromedio: number;
		volumen: number;
	};
}