// src/app/modules/ci/administracion/transacciones-configurables/model/ajusteSubSubTipo-model-interface.ts
export interface IResponseAjusteSubSubTipo {
	id: number;
	isDeleted: boolean;
	createdby: string;
	createDate: string;
	updatedby: string;
	updateDate: string;
	ajusteConfig: string;
	subsubtipo: string;
	ajusteConfigId: number;
	ajusteConfigPK: {
		id: number;
		isDeleted: boolean;
		createdby: string;
		createDate: string;
		updatedby: string;
		updateDate: string;
		codAjusteConfig: string;
		descripcion: string;
		ajusteBase: string;
		activa: boolean;
		ingreso: string;
	};
}